edk2/EdkCompatibilityPkg/Foundation/Library/Dxe/EfiDriverLib/ReportStatusCode.c:   Fix warning from Intel Compiler.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@5276 6f19259b-4bc3-4df7-8a09-765794883524

+++ EdkCompatibilityPkg/Foundation/Library/Dxe/EfiDriverLib/ReportStatusCode.c
@@ -125,7 +125,7 @@ Returns:
   EFI_STATUS                Status;
 
   DevicePathSize  = (UINT16) EfiDevicePathSize (DevicePath);
-  Size            = DevicePathSize + sizeof (EFI_STATUS_CODE_DATA);
+  Size            = (UINT16) (DevicePathSize + sizeof (EFI_STATUS_CODE_DATA));
   ExtendedData    = (EFI_STATUS_CODE_DATA *) EfiLibAllocatePool (Size);
   if (ExtendedData == NULL) {
     return EFI_OUT_OF_RESOURCES;
